# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.BulkTranslationRequest:
    properties:
      comment_ids:
        items:
          type: integer
        type: array
      target_languages:
        items:
          type: string
        type: array
    required:
    - comment_ids
    - target_languages
    type: object
...
  handlers.TranslationRequest:
    properties:
      target_languages:
        items:
          type: string
        type: array
    required:
    - target_languages
    type: object
...
  models.Article:
    properties:
      allow_comments:
        type: boolean
      author:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      author_id:
        type: integer
      blocks_version:
        description: For versioning content blocks structure
        type: integer
      bookmarks:
        items:
          $ref: '#/definitions/models.Bookmark'
        type: array
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        description: Legacy content - will be migrated to blocks
        type: string
      content_blocks:
        items:
          $ref: '#/definitions/models.ArticleContentBlock'
        type: array
      content_type:
        description: legacy, blocks, hybrid
        type: string
      created_at:
        type: string
      featured_image:
        type: string
      gallery:
        description: JSON array of image URLs
        items:
          type: string
        type: array
      has_blocks:
        description: True if article uses content blocks
        type: boolean
      id:
        type: integer
      is_breaking:
        type: boolean
      is_featured:
        type: boolean
      is_sticky:
        type: boolean
      language:
        type: string
      meta_description:
        type: string
      meta_title:
        type: string
      published_at:
        type: string
      read_time:
        description: in minutes
        type: integer
      related:
        items:
          $ref: '#/definitions/models.RelatedArticle'
        type: array
      scheduled_at:
        type: string
      slug:
        type: string
      source:
        type: string
      source_url:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        type: string
      translations:
        items:
          $ref: '#/definitions/models.ArticleTranslation'
        type: array
      updated_at:
        type: string
      user_interactions:
        items:
          $ref: '#/definitions/models.UserArticleInteraction'
        type: array
      views:
        type: integer
      votes:
        items:
          $ref: '#/definitions/models.Vote'
        type: array
    type: object
  models.ArticleContentBlock:
    properties:
      article:
        allOf:
        - $ref: '#/definitions/models.Article'
        description: Relations
      article_id:
        type: integer
      block_type:
        description: text, image, video, gallery, quote, code, divider, etc.
        type: string
      content:
        description: Main content for the block
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_visible:
        type: boolean
      position:
        description: Order within the article
        type: integer
      settings:
        description: JSON for block-specific settings
        type: object
      translations:
        items:
          $ref: '#/definitions/models.ArticleContentBlockTranslation'
        type: array
      updated_at:
        type: string
    type: object
  models.ArticleContentBlockTranslation:
    properties:
      block:
        allOf:
        - $ref: '#/definitions/models.ArticleContentBlock'
        description: Relations
      block_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      language:
        type: string
      translated_meta:
        description: For alt_text, caption, titles etc.
        type: object
      updated_at:
        type: string
    type: object
  models.ArticleTranslation:
    properties:
      article:
        allOf:
        - $ref: '#/definitions/models.Article'
        description: Relations
      article_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      language:
        description: ISO 639-1 code (en, tr, es, etc.)
        type: string
      meta_description:
        type: string
      meta_title:
        type: string
      quality:
        description: Quality score 0-1 for AI translations
        type: number
      slug:
        type: string
      status:
        description: draft, published, pending_review
        type: string
      summary:
        type: string
      title:
        type: string
      translated_by:
        description: User who created/edited the translation
        type: integer
      translation_type:
        description: manual, ai_generated, auto_translated
        type: string
      translator:
        $ref: '#/definitions/models.User'
      updated_at:
        type: string
    type: object
...
  models.Bookmark:
    properties:
      article:
        $ref: '#/definitions/models.Article'
      article_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_id:
        type: integer
    type: object
...
  models.Category:
    properties:
      articles:
        items:
          $ref: '#/definitions/models.Article'
        type: array
      children:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      color:
        description: Hex color code
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/models.Category'
        description: Relations
      parent_id:
        type: integer
      slug:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Comment:
    properties:
      article:
        allOf:
        - $ref: '#/definitions/models.Article'
        description: Relations
      article_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_edited:
        type: boolean
      parent:
        $ref: '#/definitions/models.Comment'
      parent_id:
        type: integer
      replies:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      status:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
      votes:
        items:
          $ref: '#/definitions/models.Vote'
        type: array
    type: object
...
  models.ErrorMessageTranslation:
    properties:
      category:
        description: validation, system, auth, etc.
        type: string
      created_at:
        type: string
      error_code:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      language:
        type: string
      message:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_message:
        description: User-friendly message
        type: string
    type: object
...
  models.Follow:
    properties:
      created_at:
        type: string
      follower:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      follower_id:
        type: integer
      following:
        $ref: '#/definitions/models.User'
      following_id:
        type: integer
      id:
        type: integer
    type: object
  models.FormTranslation:
    properties:
      created_at:
        type: string
      field_key:
        type: string
      form_key:
        type: string
      help_text:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      label:
        type: string
      language:
        type: string
      placeholder:
        type: string
      updated_at:
        type: string
    type: object
...
  models.LocalizedArticle:
    properties:
      allow_comments:
        type: boolean
      author:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      author_id:
        type: integer
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      content:
        type: string
      created_at:
        type: string
      featured_image:
        type: string
      gallery:
        items:
          type: string
        type: array
      id:
        description: Base article fields
        type: integer
      is_breaking:
        type: boolean
      is_featured:
        type: boolean
      is_sticky:
        type: boolean
      language:
        type: string
      meta_description:
        type: string
      meta_title:
        type: string
      original_language:
        type: string
      published_at:
        type: string
      read_time:
        type: integer
      scheduled_at:
        type: string
      slug:
        type: string
      source:
        type: string
      source_url:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      title:
        description: Localized content fields
        type: string
      translation_exists:
        description: Additional metadata
        type: boolean
      translation_quality:
        type: number
      translation_type:
        type: string
      updated_at:
        type: string
      views:
        type: integer
    type: object
  models.LocalizedComment:
    properties:
      article_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_original:
        description: Whether this is the original comment
        type: boolean
      language:
        type: string
      like_count:
        type: integer
      parent_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.LocalizedSEOSettings:
    properties:
      canonical_url:
        type: string
      keywords:
        items:
          type: string
        type: array
      og_description:
        type: string
      og_image:
        type: string
      og_title:
        type: string
      robots_follow:
        type: boolean
      robots_index:
        type: boolean
      schema:
        type: string
      twitter_card:
        type: string
      twitter_description:
        type: string
      twitter_image:
        type: string
      twitter_title:
        type: string
    type: object
...
  models.RelatedArticle:
    properties:
      article:
        allOf:
        - $ref: '#/definitions/models.Article'
        description: Relations
      article_id:
        type: integer
      id:
        type: integer
      related:
        $ref: '#/definitions/models.Article'
      related_id:
        type: integer
      score:
        description: Relevance score
        type: integer
      type:
        description: auto, manual
        type: string
    type: object
...
  models.Subscription:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      tag:
        $ref: '#/definitions/models.Tag'
      tag_id:
        type: integer
      token:
        type: string
      type:
        description: newsletter, notifications, category
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_id:
        type: integer
    type: object
...
  models.Tag:
    properties:
      articles:
        description: Relations
        items:
          $ref: '#/definitions/models.Article'
        type: array
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
    type: object
...
  models.User:
    properties:
      article_interactions:
        items:
          $ref: '#/definitions/models.UserArticleInteraction'
        type: array
      articles:
        description: Relations
        items:
          $ref: '#/definitions/models.Article'
        type: array
      avatar:
        type: string
      bio:
        type: string
      bookmarks:
        items:
          $ref: '#/definitions/models.Bookmark'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      followers:
        items:
          $ref: '#/definitions/models.Follow'
        type: array
      following:
        items:
          $ref: '#/definitions/models.Follow'
        type: array
      id:
        type: integer
      is_verified:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      location:
        type: string
      role:
        type: string
      status:
        type: string
      subscriptions:
        items:
          $ref: '#/definitions/models.Subscription'
        type: array
      updated_at:
        type: string
      username:
        type: string
      votes:
        items:
          $ref: '#/definitions/models.Vote'
        type: array
      website:
        type: string
    type: object
  models.UserArticleInteraction:
    properties:
      article:
        $ref: '#/definitions/models.Article'
      article_id:
        type: integer
      completion_rate:
        description: Percentage of article read (0.0 - 1.0)
        type: number
      created_at:
        type: string
      duration:
        description: Reading duration in seconds (for view interactions)
        type: integer
      id:
        type: integer
      interaction_type:
        description: view, bookmark, upvote, downvote, share, comment
        type: string
      ip_address:
        description: IPv4/IPv6 address
        type: string
      platform:
        description: web, mobile, app
        type: string
      referrer_url:
        description: How user found the article
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_agent:
        description: Browser/device info
        type: string
      user_id:
        type: integer
    type: object
...
  models.Vote:
    properties:
      article:
        $ref: '#/definitions/models.Article'
      article_id:
        type: integer
      comment:
        $ref: '#/definitions/models.Comment'
      comment_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      type:
        description: like, dislike
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a comprehensive CRUD API for a news blog platform with advanced
    authentication and real-time features - UPDATED!
  title: News Blog API
  version: "1.0"
paths:
...
  /api/translations/articles/{id}:
    get:
      consumes:
      - application/json
      description: Get article with localized content
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      - default: en
        description: Language code
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LocalizedArticle'
      summary: Get Article Translation
      tags:
      - translation
  /api/translations/articles/{id}/translate:
    post:
      consumes:
      - application/json
      description: Trigger article translation to target languages
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      - description: Translation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TranslationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Translate Article
      tags:
      - translation
  /api/translations/cache/invalidate:
    post:
      consumes:
      - application/json
      description: Invalidate translation cache by pattern
      parameters:
      - description: Cache pattern to invalidate
        in: query
        name: pattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Invalidate Cache
      tags:
      - translation
...
  /api/translations/comments/{id}:
    get:
      consumes:
      - application/json
      description: Get comment with localized content
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - default: en
        description: Language code
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LocalizedComment'
      summary: Get Comment Translation
      tags:
      - translation
  /api/translations/comments/{id}/translate:
    post:
      consumes:
      - application/json
      description: Trigger comment translation to target languages
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Translation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.TranslationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Translate Comment
      tags:
      - translation
  /api/translations/comments/bulk-translate:
    post:
      consumes:
      - application/json
      description: Trigger bulk comment translation to target languages
      parameters:
      - description: Bulk translation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkTranslationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Bulk Translate Comments
      tags:
      - translation
...
  /api/translations/errors/{code}:
    get:
      consumes:
      - application/json
      description: Get error message translations
      parameters:
      - description: Error code
        in: path
        name: code
        required: true
        type: string
      - default: en
        description: Language code
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ErrorMessageTranslation'
      summary: Get Error Translation
      tags:
      - translation
  /api/translations/forms/{form}/{field}:
    get:
      consumes:
      - application/json
      description: Get form field translations
      parameters:
      - description: Form key
        in: path
        name: form
        required: true
        type: string
      - description: Field key
        in: path
        name: field
        required: true
        type: string
      - default: en
        description: Language code
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FormTranslation'
      summary: Get Form Translation
      tags:
      - translation
...
  /api/translations/seo/{type}/{id}:
    get:
      consumes:
      - application/json
      description: Get SEO settings with localized content
      parameters:
      - description: Entity type (page, article)
        in: path
        name: type
        required: true
        type: string
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - default: en
        description: Language code
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LocalizedSEOSettings'
      summary: Get SEO Translation
      tags:
      - translation
...
  /api/translations/ui/{language}/{key}:
    get:
      consumes:
      - application/json
      description: Get translated UI text by key and language
      parameters:
      - description: Language code
        in: path
        name: language
        required: true
        type: string
      - description: Translation key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get UI Translation
      tags:
      - translation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"